Have created a manhatten heuristic check and added have it being called from bestFirst. Completely unchecked, but I think it looks good.

+++ IntSys_ETP/src/build1/Pair.java
@@ -8,7 +8,7 @@ public class Pair implements Comparable<Pair>
 	private double rank;
 	private LinkedList<String> route;
 	
-	Pair(double rank, LinkedList<String> route)
+	Pair(int rank, LinkedList<String> route)
 	{	this.rank = rank;
 		this.route = route; 
 	}
